cookbooks: Generalize Toolforge add node cookbook to add control nodes
Kind of a large commit but this patch adds support for adding other
types of nodes and fixes a bunch of bugs in the add worker cookbook.

Bug: T274499
Bug: T333929

+++ cookbooks/wmcs/toolforge/add_k8s_node.py
@@ -1,8 +1,9 @@
-r"""WMCS Toolforge - Add a new k8s worker node to a toolforge installation.
+r"""WMCS Toolforge - Add a new k8s node to a Toolforge cluster.
 
 Usage example:
-    cookbook wmcs.toolforge.add_k8s_worker_node \
-        --cluster toolsbeta
+    cookbook wmcs.toolforge.add_k8s_node \
+        --cluster toolsbeta \
+        --role worker
 """
 # pylint: disable=too-many-arguments
 from __future__ import annotations
@@ -34,8 +35,8 @@ from wmcs_libs.openstack.common import OpenstackServerGroupPolicy
 LOGGER = logging.getLogger(__name__)
 
 
-class ToolforgeAddK8sWorkerNode(CookbookBase):
-    """WMCS Toolforge cookbook to add a new worker node"""
+class ToolforgeAddK8sNode(CookbookBase):
+    """WMCS Toolforge cookbook to add a new node to the Kubernetes cluster"""
 
     title = __doc__
 
@@ -65,50 +66,55 @@ class ToolforgeAddK8sWorkerNode(CookbookBase):
                 "debian-10.0-buster, ex. 64351116-a53e-4a62-8866-5f0058d89c2b)"
             ),
         )
+        parser.add_argument(
+            "--role",
+            required=True,
+            choices=[role for role in ToolforgeKubernetesNodeRoleName if role.runs_kubelet],
+            type=ToolforgeKubernetesNodeRoleName.from_str,
+            default=ToolforgeKubernetesNodeRoleName.WORKER,
+            help="Role of the node to create",
+        )
 
         return parser
 
     def get_runner(self, args: argparse.Namespace) -> WMCSCookbookRunnerBase:
         """Get runner"""
-        return with_toolforge_kubernetes_cluster_opts(self.spicerack, args, ToolforgeAddK8sWorkerNodeRunner,)(
-            k8s_worker_prefix=args.k8s_worker_prefix,
-            k8s_control_prefix=args.k8s_control_prefix,
+        return with_toolforge_kubernetes_cluster_opts(self.spicerack, args, ToolforgeAddK8sNodeRunner,)(
             image=args.image,
             flavor=args.flavor,
+            role=args.role,
             spicerack=self.spicerack,
         )
 
 
-class ToolforgeAddK8sWorkerNodeRunner(WMCSCookbookRunnerBase):
-    """Runner for ToolforgeAddK8sWorkerNode"""
+class ToolforgeAddK8sNodeRunner(WMCSCookbookRunnerBase):
+    """Runner for ToolforgeAddK8sNode"""
 
     def __init__(
         self,
         common_opts: CommonOpts,
         cluster_name: ToolforgeKubernetesClusterName,
-        k8s_worker_prefix: str | None,
-        k8s_control_prefix: str | None,
         spicerack: Spicerack,
-        image: str | None = None,
-        flavor: str | None = None,
+        image: str | None,
+        flavor: str | None,
+        role: ToolforgeKubernetesNodeRoleName,
     ):
         """Init"""
         self.common_opts = common_opts
         self.cluster_name = cluster_name
-        self.k8s_worker_prefix = k8s_worker_prefix
-        self.k8s_control_prefix = k8s_control_prefix
         super().__init__(spicerack=spicerack)
         self.image = image
         self.flavor = flavor
+        self.role = role
         self.sallogger = SALLogger(
             project=common_opts.project, task_id=common_opts.task_id, dry_run=common_opts.no_dologmsg
         )
 
     def run(self) -> None:
         """Main entry point"""
-        self.sallogger.log(message="Adding a new k8s worker node")
+        self.sallogger.log(message=f"Adding a new k8s {self.role} node")
 
-        node_prefix = get_cluster_node_prefix(self.cluster_name, ToolforgeKubernetesNodeRoleName.WORKER)
+        node_prefix = get_cluster_node_prefix(self.cluster_name, self.role)
         security_group = get_cluster_security_group_name(self.cluster_name)
 
         start_args = [
@@ -134,20 +140,21 @@ class ToolforgeAddK8sWorkerNodeRunner(WMCSCookbookRunnerBase):
         ).create_instance()
         node = self.spicerack.remote().query(f"D{{{new_member.server_fqdn}}}", use_sudo=True)
 
-        device = "/dev/sdb"
-        LOGGER.info("Making sure %s is ext4, docker overlay storage needs it", device)
-        run_one_raw(
-            node=node,
-            # we have to remove the mount from fstab as the fstype will be wrong
-            command=Command(
-                f"grep '{device}.*ext4' /proc/mounts "
-                "|| { "
-                f"    sudo umount {device} 2>/dev/null; "
-                f"    sudo -i mkfs.ext4 {device}; "
-                f"    sudo sed -i -e '\\|^.*/var/lib/docker\\s.*|d' /etc/fstab; "
-                "}"
-            ),
-        )
+        if self.role.has_extra_image_storage:
+            device = "/dev/sdb"
+            LOGGER.info("Making sure %s is ext4, docker overlay storage needs it", device)
+            run_one_raw(
+                node=node,
+                # we have to remove the mount from fstab as the fstype will be wrong
+                command=Command(
+                    f"grep '{device}.*ext4' /proc/mounts "
+                    "|| { "
+                    f"    sudo umount {device} 2>/dev/null; "
+                    f"    sudo -i mkfs.ext4 {device}; "
+                    f"    sudo sed -i -e '\\|^.*/var/lib/docker\\s.*|d' /etc/fstab; "
+                    "}"
+                ),
+            )
 
         LOGGER.info("Making sure that the proper puppetmaster is setup for the new node %s", new_member.server_fqdn)
         LOGGER.info("It might fail before rebooting, will make sure it runs after too.")
@@ -160,11 +167,13 @@ class ToolforgeAddK8sWorkerNodeRunner(WMCSCookbookRunnerBase):
 
         LOGGER.info(
             (
-                "Rebooting worker node %s to make sure iptables alternatives "
+                "Rebooting %s node %s to make sure iptables alternatives "
                 "are taken into account by docker, kube-proxy and calico."
             ),
+            self.role,
             new_member.server_fqdn,
         )
+
         reboot_time = datetime.datetime.utcnow()
         node.reboot()
         node.wait_reboot_since(since=reboot_time)
@@ -175,12 +184,24 @@ class ToolforgeAddK8sWorkerNodeRunner(WMCSCookbookRunnerBase):
         )
         PuppetHosts(remote_hosts=node).run()
 
-        kubeadm = KubeadmController(remote=self.spicerack.remote(), controlling_node_fqdn=new_member.server_fqdn)
+        kubeadm = KubeadmController(remote=self.spicerack.remote(), target_node_fqdn=new_member.server_fqdn)
 
         k8s_control_node_fqdn = get_control_nodes(self.cluster_name)[0]
         kubectl = KubernetesController(remote=self.spicerack.remote(), controlling_node_fqdn=k8s_control_node_fqdn)
+        is_control = self.role == ToolforgeKubernetesNodeRoleName.CONTROL
+
+        if is_control:
+            # TODO: adjust firewall rules on etcd hosts
+            # Easiest way probably is to make the etcd profile load
+            # control nodes from PuppetDB and here just run Puppet
+            # on all of them
+
+            LOGGER.info("Copying CA data to the new server")
+            kubeadm.copy_certificates_from(existing_node_fqdn=k8s_control_node_fqdn)
 
         LOGGER.info("Joining the cluster...")
-        kubeadm.join(kubernetes_controller=kubectl, wait_for_ready=True)
+        kubeadm.join(kubernetes_controller=kubectl, wait_for_ready=True, is_control=is_control)
 
-        self.sallogger.log(message=f"Added a new k8s worker {new_member.server_fqdn} to the worker pool")
+        # TODO: for control or ingress nodes, add to the haproxy hiera key
+
+        self.sallogger.log(message=f"Added a new k8s {self.role} {new_member.server_fqdn} to the cluster")
